test file. needs a lot of work
